Use a global class for change data in gerrit-verifier-change.groovy
Metadata of the processed change is used at multiple places in the
script. The variables containing the data were handed over in function
calls. This led to functions with a large set of parameters and some
repetition. Since this data is static a global class can be used to
store this data.

Change-Id: Ib3c4e1a76af37fcd32b2add3c0fbd4237366c464