Simplify selection of build tools in gerrit-verifier-change.groovy
In gerrit-verifier-change.groovy the script decided whether Buck or
Bazel is used to build Gerrit. As it was implemented the build tools
were collected in a list, which would in theory allow multiple build
tools, but as it was implemented always only one build tool was chosen.

This change simplifies the script to assume always just one build tool
and not using lists.

Change-Id: I1b6d612b9e8d5566af086e95c2152cf4e742d2df