Remove Globals.ciTag() in gerrit-verifier-change.groovy
After a previous change (c242deb0cde3c0a0afb8eb60d69793b698f061db),
this function was only used once. In addition it caused problems,
when used in pipeline jobs. Since this script is in the progress
of being rewritten for usage in pipeline scripts, this change
removes the function.

Change-Id: I36f439af0a0bd155b55d9f9a19533d22e329812e