Add global variable to store builds in gerrit-verifier-change.groovy
In a previous change the Globals-class was renamed to Config. The
buildsList field did not logically fit into a Config-class. Also the
build tool and modes are needed by multiple functions.

This change adds a new global variable that contains the buildsList
field and fields for the build tool and modes.

Change-Id: I220cd29f1261915ec3b3317c2a78a0649161f66f