Update patch set 2
Patch Set 2:

In my opinion it might make sense to also create a GerritLabel class, since this would further separate concerns. This would also allow to get rid of some duplication in labelVerify and labelCodestyle.
If that is wanted, I would work on a change implementing that, as soon as this change is finalised.

Patch-set: 2
Reviewer: Gerrit User 1065256 <1065256@173816e5-2b9a-37c3-8a2e-48639d4f1153>