Use java.net.URLEncoder directly in gerrit-verifier-flow.groovy
In gerrit-verifier-flow.groovy the java.net.URLEncoder was added to the
String's metaclass. This was not needed and additionally did block the
conversion to a pipeline script, since changing String.metaClass is
prohibited in pipeline scripts for security reasons.

This change uses the java.net.URLEncoder directly to encode the query
URL.

Change-Id: Iaa5ada7c1c79e7e95f6ed91ecc636b07ababefb6