Configure mvn- and sbt-slave usage
The job-declarations use the node label mvn and sbt to build mvn- and
sbt-dependent builds exclusively in the respective containers. The
config.xml did not yet contain the respective labels. These jobs could
thus not find the correct slave containers.

This change adds the labels to the configuration of the mvn- and
sbt-slave and configures them to be exclusively built by jobs requesting
a node with the respective label.

Change-Id: Iafa647df1f4a8d79c065720fdf6c1341a4cf1025