Update patch set 4
Patch Set 4:

After using docker-plugin 1.0.4 for a few days with Jenkins 2.25, I noticed that there seems to be some problem with the docker-plugin. The docker containers start up, but are sometimes stopped immediately without error, leading to inefficient execution of jobs. I am using the SSH-injection method to connect to the docker containers (needed for our setup), so this might be part of the problem. Since gerrit-ci.gerritforge.com is using docker-plugin 1.0.4 with Jenkins 2.25 and connection with build-in ssh-keys, it may indeed be the case that SSH-injection is still buggy in that plugin version. Updating to Jenkins 2.60.3 and docker-plugin 1.1.4 was enough to solve the problem.
If I find the solution, I will implement it here, but I would suggest to use this change series only together with the update to 2.60.3, to avoid introducing problems in the stable setup.

Patch-set: 4
Reviewer: Thomas Dräbing <1065256@173816e5-2b9a-37c3-8a2e-48639d4f1153>