Configure Jenkins to use debian-jessie based bazel-slave
Previous changes changed the base docker images for the build slaves to
use debian jessie. Jenkins was not yet configured to use these new images.

This change adjusts the configuration to use the
gerritforge/gerrit-ci-slave-bazel:debian-jessie image for bazel builds
and since it is now the only bazel slave in use, it simplifies the label
used for this slave to bazel instead of bazel-debian.

Change-Id: I03d3f468bd5fddfbf097c457e3cdcaa037db9463