Adapt output parsing of google-java-format
The newer versions of google-java-format is not issuing "Need
Formatting" output prior to dumping the actual files:

  ... | xargs -r ~/format/google-java-format-1.6 -n --set-exit-if-changed
  gerrit-gwtui/src/main/java/com/google/gerrit/client/FormatUtil.java

Given that we parse the output of multiple commands, we need to re-add
some indication that we are parsing the output of gjf. Redirect the
output to a file and evaluate the return value of the gjf command. In
case it is not equals zero, re-add "Need Formatting" in the beginning and
open and close brackets to the output around the list of formatted files.

Bug: Issue 8829
Change-Id: I958a307f25d15a1fbdcaa613ee21e4173d69485d