Fix Makefile of debian- and chrome-slaves to work for both debian versions
The Makefile for slave-debian and slave-chrome contains the targets
'build', 'publish' and 'clean', which are supposed to build the image
with debian stretch and debian jessie as inherited image respectively.
This did not work correctly, because the docker_build and docker_publish
target were called twice. Make does not execute the same target twice and
thus only one of both images was build and published. The clean target
was not yet adapted to work with multiple debian versions.

This change fixes the 'build', 'publish' and 'clean'-targets to
correctly work with both debian versions.

Change-Id: I3267cd3b434a0a31ea21b4a168f017e8ad7dc2e1