Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> What about fixing the plugin instead?

It's not only a case of fixing the plugin; the elasticsearch-metrics-reporter-java project also needs to be fixed. This is not likely to be a trivial task, so I'd prefer to stop building this plugin to prevent anyone expecting it to work.

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>