Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review+2
> 
> > Patch Set 4:
> > 
> > > Patch Set 2: Code-Review-1
> > > 
> > > +1 to disabling ReviewDb tests on master, but we still need to not break ReviewDb on 2.16 when we make changes on the stable-2.16 branch.
> > 
> > Thanks for the feedback: will disable then only on master and will update the 2.16 release notes saying that ReviewDb is still supported but *migration to NoteDb is mandatory* for being able to upgrade to 3.0.
> 
> Yep, thanks :). But the fact that they have to migrate at some point after 2.16 means they can be running 2.16 with ReviewDb for some amount of time beforehand.

True, that's why I agree that we need to keep on running the ReviewDb tests on 2.16, but NOT on 3.0 (master).

What are you planning to do with 3.0? Simply refuse to run Gerrit if on ReviewDb? Will off-line migration to NoteDb still supported?

Patch-set: 5
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>