Revert "Remove outdated docker-plugin Snapshot build"
This reverts commit 1543664f25e3cdeb9cfe5cf8e03c55a0cb64d183.

Reason for revert: I did not test this properly before and,
once the machine was entirely rebuilt, this version of the
plugin broke the slave provisioning system. Rolling back this
change, it works again.

I would suggest to revert it for now and then build a brand-new
Gerrit-CI system for testing all changes like this one.

Change-Id: I693207d9b03dfdacc0a30d82888e9eb5a1704da5