Update patch set 3
Patch Set 3: Code-Review+1

Sorry, that my change broke the system. Creating a test-instance for critical changes would definitely be a good idea.

In our fork this change was working, but we also have the related change 185022 implemented. I made this change, because the old ubuntu-based slave couldn't be built anymore and thus your method of providing SSL-certificates would not work for us, since the certificates build into the image did not have the same CA. This change of the config switches the way of how Jenkins communicates with the slaves. With the change Jenkins injects the SSH-keys into the slave-container.

Since checking the configuration of the Gerrit-CI showed that a newer docker plugin version than 0.16.2 was installed (the plugins.txt also sets a newer version), I didn't think that would be an issue, but maybe remnants of the old 0.16.2-version were still being used.

Anyway, I agree that we should only continue looking into it, when a testing instance is available. It is hard for me to verify changes submitted here properly, since our fork now already includes some changes that were necessary for our particular needs.

Patch-set: 3
Reviewer: Gerrit User 1065256 <1065256@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+1