Update patch set 1
Patch Set 1:

I'm not 100% sure we want this yet.

If we stop running CI against ReviewDb, then it's only a matter of time until something breaks--and I'm willing to bet the amount of time will be short. In that light, this change is tantamount to saying it is ok to start breaking/removing ReviewDb code. I wonder if it's too soon to start saying that.

So my inclination is to say, let's talk about this at the hackathon next week, where I plan to actively work on NoteDb schema upgrades.

On the other hand, from a tactical perspective, submitting this now will definitely increase CI throughput at the hackathon, so maybe it's a good idea for that reason alone. I could be convinced.

Patch-set: 1
Reviewer: Gerrit User 1010008 <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>