Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'm not 100% sure we want this yet.
> 
> If we stop running CI against ReviewDb, then it's only a matter of time until something breaks--and I'm willing to bet the amount of time will be short.

Yes, but I believe that leaving the validation and the tests inside the Gerrit code-base would mean that we will need to keep supporting all those use-cases, which would slow us down.

> In that light, this change is tantamount to saying it is ok to start breaking/removing ReviewDb code. I wonder if it's too soon to start saying that.

*IF* if we are not close to removing any ReviewDb code as yet, then we should keep the validation. On the other side if we want to get rid of it ASAP on master, to make our code cleaner, build faster and tests simpler, then we should start first removing the validation on the CI.

> So my inclination is to say, let's talk about this at the hackathon next week, where I plan to actively work on NoteDb schema upgrades.
> 
> On the other hand, from a tactical perspective, submitting this now will definitely increase CI throughput at the hackathon, so maybe it's a good idea for that reason alone. I could be convinced.

Han-Wen gave me extra power, so it is not a problem for the hackathon. It could be possibly *after* the hackathon when the power will be gone and we'll be back with only 2 docker servers :-(

Patch-set: 1
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>