Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Running now a stress test of the gerrit-ci with the rollback (2nd) of the docker-plugin. If that is good, I'll verify and merge.
> 
> @Thomas there is a lot more work to do on the docker-plugin, even the master version is broken :-(

I see. Is it keeping us from implementing the tests/test instances for gerrit-ci-scripts?

In our fork we are currently using the Inject SSH key method (as I proposed in an earlier change), which works well for us, but we have a much lower load of course. The attach-method seems interesting and it would be nice to consolidate our fork with the original project in that regard again :-), but let's wait until these issues are resolved.

Patch-set: 2