Create multibranch-pipieline-job to verify Gerrit
The current implementation of jobs verifying changes in the
gerrit-project uses the deprecated build-flow plugin. This keeps us from
updating to newer versions of Jenkins. These jobs also use a lot of
custom code to retrieve changes from Gerrit and to apply labels, which
would not be needed using a multibranch pipeline in combination with the
gerrit-codereview-plugin for Jenkins.

This change adds a multibranch-pipeline-job to the Gerrit-CI that uses
a Jenkinsfile (see corresponding change 204010 in the gerrit-project) to
verify changes on Gerrit.

TODOs:
 - Finish implementation of Jenkinsfile
 - The old verification jobs were left intact for now to not break
   verification during testing. These should either be removed in
   a later patchset or in a separate change.

Change-Id: I09f82e7e7a498cab82be50793ae82c9d18aa4252