Allow to build specific jobs with gerrit-ci-scripts-manual
The gerrit-ci-scripts-manual job always built all jobs and deleted all
old jobs. Sometimes, e.g. when testing a Jenkins instance, it might be
useful to just create specific jobs.

This change adds an additional parameter to the gerrit-ci-scripts-manual
job that allows to specify one or multiple jobs to be created/updated.
If this parameter is left blank (default) all jobs will be created.

Change-Id: I007504709dff71954a57954070d3f4b884a249f9