Update patch set 2
Patch Set 2:

We are not using the groovy scripts on our instance anymore. (We switched to Gerrit trigger for simplicity.)
I could try to create the jobs manually in our Jenkins next week and test the change manually, if that would be enough to give it 'Verified +1'. But a better approach would probably be to use the test instance Luca set up this week to build some patchsets, where we know that the google-java-format is complaining. Luca, do you have time to do that?

Patch-set: 2
Reviewer: Gerrit User 1065256 <1065256@173816e5-2b9a-37c3-8a2e-48639d4f1153>