Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3: Verified-1
> > 
> > Did not work (see https://gerrit-review.googlesource.com/c/gerrit/+/204935) and produced an NPE on the script (see https://gerrit-ci.gerritforge.com/job/Gerrit-verifier-change/34914/console)
> 
> How can I troubleshoot it? How you debug that kind of problems?

Jenkins workflow-scripts are hell to debug. The only possibility is to my knowledge is to run the Job in Jenkins again and again, making changes and adding debug output, until it works. I always created a copy of the jobs to test, to not disrupt the running jobs. If somebody knows a better solution, please tell.

Patch-set: 3
Reviewer: Gerrit User 1065256 <1065256@173816e5-2b9a-37c3-8a2e-48639d4f1153>