Do not create a new ssh-key with each 'make build' of master
Each time 'make build' and therefore also 'make start' was used to
build/start the master image a new ssh key was created and built into
the docker image, if make's cache was emptied in the meantime for some
reason. This then required to rebuild all slave images, since the new
key was not saved in the authorized_keys-file of the slave images.
This created a huge overhead, when testing changes on the Gerrit
master.

This change adds as a condition to create a new key, that the private or
public key is missing or if caching of the keys is disabled (default:
enabled).

Change-Id: Ic7cee2376bca22075ec335fd9406f21def811ea0