Remove custom entrypoint from bazel-docker node
In the Jenkins configuration for the docker slave running bazel a custom
entrypoint was configured. This entrypoint was not compatible with the
slave image as present in this project and thus builds would not start.

This change removes this configuration from the config, which allows
Jenkins to connect to the slave-bazel image as it is defined in this
project.

Change-Id: Ia11c4009c4a8e099eb986560135808b3e195b94a