Remove DNS and custom entrypoint from bazel-docker node
In the jenkins configuration for the docker slave running bazel a custom
DNS and entrypoint were configured. These were not compatible with the
salve image as present in this project and thus builds would not start.

This change removes this configuration from the config, which allows
Jenkins to connect to the slave-bazel image as it is defined in this
project.

Change-Id: Ia11c4009c4a8e099eb986560135808b3e195b94a