Set root as the default user in slave-mvn
The docker-plugin of Jenkins requires the default user of a docker
container to be root, to be able to connect via SSH, even though all
operations in the container concerning builds are run as the user
jenkins. This was not the case for slave-mvn. Thus the connection
to the slave-mvn containers failed.

Change-Id: I65c8a6a3fa95aba784461b33099f1dd08ce9e94b