Update patch set 23
Patch Set 23:

> Patch Set 23:
> 
> > Patch Set 23: Verified+1
> > 
> > Tested on my local machine: Builds (make build) without error; Jenkins starts (make start) up successfully; the gerrit-ci-scripts job runs without errors.
> 
> Since this upgrade, the jobs are not removed anymore during update :-(
> Do you know how to fix the problem?

Could it be that the change 204611 is causing it? There, I head to remove the --delete-old flag, because otherwise, all jobs that are not currently built, when just building one job, would be removed, which would of course be not desirable. The gerrit-ci-scripts job should still remove old (unmanaged) jobs.
Extending the script to look whether a selection of jobs should be build, and if not delete unmanaged jobs on update should be an easy fix.

Patch-set: 23