Delete unmanaged jobs with gerrit-ci-scripts-manual
After change 204611 unmanaged jobs were not deleted anymore, when
running gerrit-ci-scripts-manual. This was intentional, since doing so
would cause all other jobs to be deleted, if creating just a specific
job. Nevertheless, this is inconvenient since unmanaged jobs could
accumulate.

This change implements a check, whether only specific jobs are supposed
to be created/updated and if that is not the case deletes old unmanaged
jobs.

Change-Id: I7261e5fa91dae2f3080f41effcd4c41f4f7de74b