Update patch set 1
Patch Set 1:

I definitely think, we should reduce the size of the images. 
We also already discussed in our team, how the caching could be handled a bit better. The idea that came up was to maintain a separate Docker volume, containing the cached dependencies, that is mounted by the containers. That would move the caches out of the docker images and they would also be constantly be up-to-date, since the builds would download new dependencies. A big issue would however be, when two builds would download a new dependency at the same time. I don't know whether buck and bazel have safeguards for cases like this. So I am not sure yet, whether that would work well.

Patch-set: 1
CC: Gerrit User 1065256 <1065256@173816e5-2b9a-37c3-8a2e-48639d4f1153>