Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I definitely think, we should reduce the size of the images. 
> 
> Indeed :-) This is the first step, getting rid of stuff that we do not use, *for sure*
> 
> > We also already discussed in our team, how the caching could be handled a bit better. The idea that came up was to maintain a separate Docker volume, containing the cached dependencies, that is mounted by the containers. That would move the caches out of the docker images and they would also be constantly be up-to-date, since the builds would download new dependencies. A big issue would however be, when two builds would download a new dependency at the same time. I don't know whether buck and bazel have safeguards for cases like this. So I am not sure yet, whether that would work well.
> 
> In your case you could manage with a mounted volume, but for gerrit-ci, I do not have access to anything on the remote Docker hosts, other than the dockerd APIs.
> 
> I am open to creative ideas :-)

I was thinking about using `docker volume create` [1] for this purpose. This would use the docker API to create a volume. This would then be a named volume that could be mounted into the slave containers. If I understood it correctly no direct access to the filesystem is needed, docker will create the volume in its data directory. However, I guess that would have to be done on every docker node and caches could differ.

[1] https://docs.docker.com/engine/reference/commandline/volume_create/

Patch-set: 1