Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > I definitely think, we should reduce the size of the images. 
> > 
> > Indeed :-) This is the first step, getting rid of stuff that we do not use, *for sure*
> > 
> > > We also already discussed in our team, how the caching could be handled a bit better. The idea that came up was to maintain a separate Docker volume, containing the cached dependencies, that is mounted by the containers. That would move the caches out of the docker images and they would also be constantly be up-to-date, since the builds would download new dependencies. A big issue would however be, when two builds would download a new dependency at the same time. I don't know whether buck and bazel have safeguards for cases like this. So I am not sure yet, whether that would work well.
> > 
> > In your case you could manage with a mounted volume, but for gerrit-ci, I do not have access to anything on the remote Docker hosts, other than the dockerd APIs.
> > 
> > I am open to creative ideas :-)
> 
> I was thinking about using `docker volume create` [1] for this purpose. This would use the docker API to create a volume. This would then be a named volume that could be mounted into the slave containers. If I understood it correctly no direct access to the filesystem is needed, docker will create the volume in its data directory. However, I guess that would have to be done on every docker node and caches could differ.

Exactly, that's the problem. If I create the volume, let's say, in the Gerrit-verify-cache groovy script, I would create it on the Gerrit master (where the script gets executed) and the actual build could run in any of the slaves around the globe.

This is the unsolved problem of the distributed volumes in Docker ... still unsolved in Kubernetes as well.

The second problem is: if I start re-using a volume across builds, they become all of a sudden not reproducible. That means build-B done after build-A will reuse cached information instead of resolving it again from a stable consistent point.

I am in favor of speeding up with a "well known" cache, but not a read/write volume shared across containers. 

What about, instead, a read-only volume? But then the problem is that the Bazel build won't be able to add additional dependencies on it.

Patch-set: 1
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>