Update patch set 2
Patch Set 2:

(3 comments)

Another thing I noticed is, that the debian-stretch based slaves are a bit smaller than the debian-jessie based slaves (~100Mb for slave-chrome). I understand that we can't easily get rid of debian-jessie yet, but could we already build a stretch based, slave-buck/slave-bazel and use labels in Jenkins to point jobs that still need Java 7 to the jessie based, which should not happen that often. For that we would need another condition in the slave-buck to only install buck for java 7, if debian-jessie is used. That would also be a more step-wise approach of getting rid of debian-jessie.

Patch-set: 2