Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Why it still works? E.g.: [1], both modes are tested.
> 
> [1] https://gerrit-review.googlesource.com/c/gerrit/+/208712

No, the notedb tests are not really running. The job starts, but as part of the job the gerrit-bazel-verifier-test.sh tests, whether the string given as a parameter is equal to *"notedb"* (line 23). This test is negative (you can see that in the console log of the test). The script ends successfully (thus the reported SUCCESS), but without actually running the tests.

Patch-set: 1
Reviewer: Gerrit User 1065256 <1065256@173816e5-2b9a-37c3-8a2e-48639d4f1153>