Use slave-bazel as default slave
With change 207914 the slave-buck image was removed from the Jenkins
configuration as a docker agent. Since the slave-buck agent was the
default and all other images are set to be used only by jobs specifying
the label of a build agent, a lot of jobs would not be able to be built
with the current configuration, since they did not specify a label and
thus required a default.

This change makes the slave-bazel agent the default. Since the container
image contains buck as well, it can serve buck- as well as bazel-builds
and thus is able to serve all jobs expecting a default.

Change-Id: I3a3a20a213fc487f76eb58aefea69e42e2404831