Update patch set 4
Patch Set 4:

I believe that at the moment we're a bit stuck on the approach: shall we keep updating the single Docker image for all branches and thus updating all the BUILD for all of them?

Or should we give up embedding the Bazel setup in the Dockerfile and dynamically install the one with the version needed at the beginning of the build script?

WDYT?

Patch-set: 4