Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I believe that at the moment we're a bit stuck on the approach: shall we keep updating the single Docker image for all branches and thus updating all the BUILD for all of them?
> 
> Or should we give up embedding the Bazel setup in the Dockerfile and dynamically install the one with the version needed at the beginning of the build script?
> 
> WDYT?

For now, we indeed depend not only on Bazel installed locally and thus the right version üer branch, but also the bazel-cache that is consumed from the docker image. We build :release and other targets on all 4 branches: 2.14, 2.15, 2.16 and master during the docker image build and therefore the CI-build is a no-op, because the cache is warmed.

Strictly speaking even if we don't bump Bazel version we should at least monthly (or even more often?) rebuild docker Bazel-image for just that: refresh the Bazel build cache with the tip of the stable branches and master to reduce the build CI-time.

If we stop pinning Bazel in docker image, we would immediately lose the caching and the build would take much longer.

Luckily, Bazel supports remote cache and it advanced a lot in recent versions and becoming more mature and it is used on Bazel team's own CI too.

IOW the short term solution is to keep docker-image/Bazel-version+Bazel-cache pinning. Mid/long term solution is to de-couple docker image from Bazel version+Bazel cache and move to using remote Bazel Cache. As immediate consequence of that move we would have to set-up dedicated Bazel cache docker image with a lot of storage (can someone from Gerrit community donate it?) and periodically wipe it out.

Patch-set: 4
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>