Update patch set 2
Patch Set 2: Verified+1

I noticed that on gerrit-ci.gerritforge.com builds using Buck actually were already built with the bazel slave, which would not work with the EXCLUSIVE-mode (Only build jobs with label expressions matching this node), that only allows jobs with the label of the node configured to it, to be built by the node. I quickly checked with Sasa, how the configuration in Jenkins looked like and saw that it was already changed to NORMAL-mode (Use this node as much as possible).
@Luca, could you have a look at this change and submit it, if it is okay with you?

Patch-set: 2
Reviewer: Gerrit User 1065256 <1065256@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=+1