Update patch set 2
Patch Set 2: -Verified

> Patch Set 2: Code-Review+1
> 
> Hopefully this upgrade wouldn't cause regression with recent Jenkins docker plugin upgrade, that was reverted last time.

Yes, we should avoid that. But the git-client plugin not working is also no solution, since we then can't use the Git SCM functionality of Jenkins. That was at least what caused me to propose this change.

As long as we don't rebuild the master-container and reinstall the plugins by starting the new image, we won't run into this issue, but that is certainly no long term solution.

In case that this combination does not work, we would have to find out, which dependency versions have to be fixed for the current plugin versions to work.

The dependency chain is as follows:

docker-plugin -> ssh-slaves -> ssh-credentials -> credentials

So this change effects at least an indirect dependency.

Patch-set: 2
Reviewer: Gerrit User 1065256 <1065256@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0