Fix naming of DOCKER_HOST variable in master docker file
The docker host url was set in the config.xml file using xsltproc using
the env variable DOCKER_HOST. In the Dockerfile of the master container
this variable was wrongly named DOCKER_URL. This could cause xsltproc to
fail and thus to fail to create the config.xml.

This change fixes the variable name.

Change-Id: I7db0ff37f8ff1a6249dd9b963bdd19f888c2c859