Update patch set 1
Patch Set 1: Verified+1 Code-Review+2

> Patch Set 1:
> 
> > If we come into a situation that one branch can't be built anymore with a newer version of bazel, we should maintain multiple versions of the bazel-slave. 
> 
> I disagree with this assessment. This is a clear violation of YAGNI principle: never do something in advance in case something could ever happen. Not in your code base, and not in your infrastructure!
> 

If there is no need to maintain multiple bazel build slaves, we should not maintain them. I understood Luca's comment like that there were issues in that direction.

> We started with Bazel 0.6 3 years ago. Today we are on Bazel 0.17.2/0.22.0. All tips of 3 stable branches managed by Bazel can be built today with the latest Bazel releases. Moreover I'm checking weekly that Gerrit and the whole ecosystem (JGit, Gitiles and Gerrit plugins that I maintain) can be built with Bazel@HEAD to ensure we don't have any Bazel official releases that cannot built Gerrit. Check this release thread on Bazel issue tracker, to see what I mean: [1].
> 
> In fact I dscovered these breakages on Bazel@HEAD and they were all fixed before the Bazel versions were released: [2], [3].

That is good to know :-). Thanks for your work on that!

> 
> Besides that, how is this discussion related to this change? This change does not change anything at all. It's a good refactoring for 2 variables.
> 

A discussion like this ("Do we need to support multiple bazel-versions at the same time.") came also up with previous changes touching the bazel-version. This question should be answered, but yes, it is not directly related to this change and should better be discussed elsewhere.

> If someone thinking is, we really need explosion of Docker images pro Debian/Bazel version, I wouldn't object against that change.
> 
> * [1] https://github.com/bazelbuild/bazel/issues/6493
> * [3] https://github.com/bazelbuild/bazel/issues/7191
> * [4] https://github.com/bazelbuild/bazel/issues/7102

So lets submit this change and if needed create a separate change to introduce the new naming-scheme suggested by Luca.

Patch-set: 1
Reviewer: Gerrit User 1065256 <1065256@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=+1
Label: Code-Review=+2