Pin version of eslint-plugin-html to 5.0.3
The eslint-plugin-html package version was not pinned. This could lead
to issues, if a change was made to this project that does not conform
with the currently used codestyle of the Gerrit project, leading to
failing codestyle verifications.

Changing the adapted codestyle rules for the Gerrit project should be
a planned step, thus this change pins the version of eslint-plugin-html
to the current version, making an update a deliberate choice.

Change-Id: I109456a7f116f67c860786b664483b8c4c258c9b