Update patch set 2
Patch Set 2:

The build passes for me locally.  The tests are known to fail unless the postgres environment is configured properly; this should also be the case for the existing branches stable-2.14 and stable-2.15 so CI should not run tests for this plugin.

http://paste.openstack.org/show/744989/

Patch-set: 2
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>