Build slave-mvn after slave-bazel
When using the Makefile in ./jenkins-docker, the slave-mvn image was
built before the slave-bazel image, which it inherits. Thus, the
slave-mvn image was based on an outdated slave-bazel image, when being
built using this method.

This change moves the position of slave-mvn in the build-queue defined
in the Makefile after slave-bazel.

Change-Id: I18025c6c0899c30fa526de56be7d7ab38493ed27