Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Running docker in docker does not seem advisable [1]. 
> 
> This is what we have been doing for years on the Gerrit-ci and it works.
> 
> We basically use the external docker daemon through the DOCKER_HOST variable.
> 
> > I would rather suggest, to mount the docker.sock, which we already mount to the master-container, also to the bazel slave or make the worker nodes' docker usable in another way. That would be more stable.
> 
> This is what we do.
> 
> > I am not sure though, whether this is an option with the current production-setup using remote worker nodes.
> 
> Yes, it is.
> 
> > 
> > [1] https://stackoverflow.com/questions/27879713/is-it-ok-to-run-docker-from-inside-docker

I of course know that we are using docker from a docker-container. This change made it look as if you wanted to use the docker.sock installed in the container by docker-ce and not mount the docker.sock from the host. That was the reason for my concern.

As David Ostrovsky wrote the image is not used anywhere. We should add it to the master's config.xml and also add the volume mount for the docker.sock to its config.

Patch-set: 1
Reviewer: Gerrit User 1065256 <1065256@173816e5-2b9a-37c3-8a2e-48639d4f1153>