Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 6: Code-Review+1
> > 
> > All the docker images are updated on the CI and can run GJF 1.7: shall we enable validation also using the new specs?
> > 
> > WARNING: All the validations would fail, unless they use GJF 1.7. We should communicate that publicly.
> 
> This change formats with 1.7 for stable-2.14 but is still failing.  How can we make it use 1.7?
> 
> https://gerrit-review.googlesource.com/c/gerrit/+/211932

The script is hardcoded to use GJF 1.6:

  xargs -r /home/jenkins/format/google-java-format-1.6 -n --set-exit-if-changed

I guess, we could add an if-clause to test for which branch to use which version. Then we could add new branches to test with GJF 1.7 as soon as they are ready.

Or we could try to use the tools/setup-gjf.sh script in the Gerrit-codestyle job to dynamically set the GJF version.

What do you think?

Patch-set: 6
CC: Gerrit User 1065256 <1065256@173816e5-2b9a-37c3-8a2e-48639d4f1153>