Use setup-gjf.sh script to get required gjf version
The Gerrit-codestyle-job had the Google-Java-Formatter version
hard-coded in the script. Thus, if the version was changed in Gerrit, it
had to be changed simultaneously in the CI and for all Gerrit-branches.

This change changes the Gerrit-codestyle job to use the
./tools/setup-gjf.sh-script to get the version of GJF configured for
Gerrit in the current commit. This allows to have different GJF-versions
between Gerrit-branches and removes the requirement to change the
GJF-version in the CI.

Change-Id: Ie5b5cf8226cb8461b6639b076a0ef57ae8b8afd9