Avoid downloading existing versions of Google-Java-Format
The setup_gjf.sh-script downloads Google-Java-Format (gjf) each time it
is run. This caused a download each time a codestyle test was executed.

This change adds a test, whether the gjf-version set as default in
setup_gjf.sh is already present in the build slave container. If it is
present, the executable built into the container will be used, else the
setup_gjf.sh will be run to install the required version of gjf.

Change-Id: If4eb9e3e2a3ed2f7062c18db7a974c903d71c27e