Update patch set 5
Patch Set 5: -Verified

> Patch Set 5:
> 
> > Patch Set 5: Verified-1
> > 
> > Unfortunately, the jobs for both branches still fail, because of test failures:
> > 
> > master: http://paste.openstack.org/show/Q0cNlkEYo183t0cZIr5W/
> > stable-2.16: http://paste.openstack.org/show/TcLu614cS03vp8k5KM13/
> > 
> > Do the tests work for you, when run locally?
> 
> It passes for me locally.  Can you check what exactly is failing?

I think the issue might be, that the test system did not use the slave-bazel-docker image for the jobs, since it is not configured by default yet. I would have to build the image on the test system first to test that, but unfortunately due to the stop of support for debian jessie (see my email on the mailing list) the container images do not built at the moment. Thus, I am unable to verify this change right now.

Patch-set: 5
Reviewer: Gerrit User 1065256 <1065256@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0